pyproject.toml
[project]
name = "rehearten"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.0.1",
    "beautifulsoup4>=4.12.0",
    "certifi>=2023.7.22",
    "cryptography>=41.0.0",
    "django>=4.2.0,<5.0",
    "django-cors-headers>=4.0.0",
    "django-debug-toolbar>=4.1.0",
    "django-social-auth>=0.7.28",
    "djangorestframework>=3.14.0",
    "markdown>=3.4.0",
    "mongoengine>=0.27.0",
    "openai>=1.0.0",
    "pillow>=10.0.0",
    "pymongo>=4.3.0",
    "python-dateutil>=2.8.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "requests-oauthlib>=1.3.0",
    "social-auth-app-django>=5.2.0",
    "social-auth-core>=4.4.0",
    "urllib3>=2.0.0",
]
